pyproject.toml
[project]
name = "meme-convention"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "keyboard>=0.13.5",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pyperclipimg>=0.2.0",
    "pytest>=8.4.1",
]
